Create general connection manager class
